+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.0 #
 
+* [!] `destroy()` accepts one or more elements on which the changes applied by the library have to removed
 * Code refactored
 
 # 0.1.0 (13-10-2014) #